Move hooks each folders that have related
